Added error handling and more logging

+++ Syncify/MP3Processor.cs
@@ -9,32 +9,49 @@
 
     public static class MP3Processor
     {
-        public static void RemoveImagesFromMP3Files(string directoryName)
+        public static void RemoveImagesFromMP3Files(string directoryName, ILogger logger)
         {
-            RemoveImagesFromDirectory(directoryName);
+            RemoveImagesFromDirectory(directoryName, logger);
 
-            RemoveImageFromIDTag(directoryName);
+            RemoveImageFromIDTag(directoryName, logger);
         }
 
         public static void RetitleMP3Files(string directoryName, ILogger logger)
         {
-            var mp3FilePaths = GetMP3FilesInDirectory(directoryName);
-
-            foreach (var currentFile in mp3FilePaths)
+            if (logger != null)
             {
-                using (var file = TagLib.File.Create(currentFile))
+                try
                 {
-                    if (!RenamedAlready(file.Tag.Title))
+                    if (Directory.Exists(directoryName))
                     {
-                        var newTitle = file.Tag.Track.ToString("00") + " " + file.Tag.Title;
-                        var oldTitle = file.Tag.Title;
+                        var mp3FilePaths = GetMP3FilesInDirectory(directoryName);
 
-                        file.Tag.Title = newTitle;
+                        foreach (var currentFile in mp3FilePaths)
+                        {
+                            using (var file = TagLib.File.Create(currentFile))
+                            {
+                                if (!RenamedAlready(file.Tag.Title))
+                                {
+                                    var newTitle = file.Tag.Track.ToString("00") + " " + file.Tag.Title;
+                                    var oldTitle = file.Tag.Title;
 
-                        logger.LogInfo(oldTitle + " renamed " + newTitle);
+                                    file.Tag.Title = newTitle;
+
+                                    file.Save();
+
+                                    logger.LogInfo(oldTitle + " retitled " + newTitle);
+                                }
+                                else
+                                {
+                                    logger.LogWarning(file.Tag.Title + " could not be retitled");
+                                }
+                            }
+                        }
                     }
-
-                    file.Save();
+                }
+                catch (Exception ex)
+                {
+                    logger.LogError(ex.Message);
                 }
             }
         }
@@ -49,30 +66,60 @@
             return Directory.EnumerateFiles(directoryName, "*.mp3", SearchOption.AllDirectories);
         }
 
-        public static void RemoveImageFromIDTag(string directoryName)
+        public static void RemoveImageFromIDTag(string directoryName, ILogger logger)
         {
-            var mp3FilePaths = GetMP3FilesInDirectory(directoryName);
-
-            foreach (var currentFile in mp3FilePaths)
+            if (logger != null)
             {
-                using (var file = TagLib.File.Create(currentFile))
+                try
                 {
-                    file.Tag.Pictures = Array.Empty<TagLib.IPicture>();
+                    if (Directory.Exists(directoryName))
+                    {
+                        var mp3FilePaths = GetMP3FilesInDirectory(directoryName);
 
-                    file.Save();
+                        foreach (var currentFile in mp3FilePaths)
+                        {
+                            using (var file = TagLib.File.Create(currentFile))
+                            {
+                                file.Tag.Pictures = Array.Empty<TagLib.IPicture>();
+
+                                file.Save();
+
+                                logger.LogInfo("Image removed from " + file.Tag.Title);
+                            }
+                        }
+                    }
+                }
+                catch (Exception ex)
+                {
+                    logger.LogError(ex.Message);
                 }
             }
         }
 
-        public static void RemoveImagesFromDirectory(string directoryName)
+        public static void RemoveImagesFromDirectory(string directoryName, ILogger logger)
         {
-            var jpgFiles = GetJpgFilesInDirectory(directoryName);
-
-            if (jpgFiles.Any())
+            if (logger != null)
             {
-                foreach (var jpg in jpgFiles)
+                try
                 {
-                    File.Delete(jpg);
+                    if (Directory.Exists(directoryName))
+                    {
+                        var jpgFiles = GetJpgFilesInDirectory(directoryName);
+
+                        if (jpgFiles.Any())
+                        {
+                            foreach (var jpg in jpgFiles)
+                            {
+                                File.Delete(jpg);
+                            }
+
+                            logger.LogWarning("Images removed from " + directoryName);
+                        }
+                    }
+                }
+                catch (Exception ex)
+                {
+                    logger.LogError(ex.Message);
                 }
             }
         }
